server/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Cors==4.0.1
python-dotenv==1.0.1
werkzeug==3.0.3
pytest==8.3.2
pytest-cov==5.0.0